# ticket_management_system/swagger_specs/swagger.yml
swagger: '2.0'
info:
  title: Flight Management System API
  description: RESTful API for managing flights and bookings with scoped user-ID tokens
  version: 1.0.0
basePath: /api
host: localhost:8080
schemes:
- http
- https
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      JWT Authorization header using the Bearer scheme.
      Example: "Authorization: Bearer {token}"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    default: 1
    description: Page number (1-indexed)
  PerPageParam:
    name: per_page
    in: query
    type: integer
    default: 10
    description: Number of items per page (max 100)
...
definitions:
...
  ValidationError:
    type: object
    properties:
      error:
        type: string
        example: Bad Request
      message:
        type: string
        example: Validation failed
      errors:
        type: object
        additionalProperties: true
        example:
          email:
          - Not a valid email address
          password:
          - Password must be at least 6 characters
    required:
    - error
    - message
  AuthError:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Invalid or missing authentication token
    required:
    - error
    - message
  NotFoundError:
    type: object
    properties:
      error:
        type: string
        example: Not Found
      message:
        type: string
        example: Resource not found
    required:
    - error
    - message
  ConflictError:
    type: object
    properties:
      error:
        type: string
        example: Conflict
      message:
        type: string
        example: Resource already exists or operation cannot be performed
    required:
    - error
    - message
  ForbiddenError:
    type: object
    properties:
      error:
        type: string
        example: Forbidden
      message:
        type: string
        example: You do not have permission to access this resource
    required:
    - error
    - message
  User:
    type: object
    properties:
      id:
        type: string
        format: uuid
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      role:
        type: string
        enum:
        - user
        - admin
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  UserResponse:
    type: object
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/User'
      token:
        type: string
        description: Scoped JWT token
      token_type:
        type: string
        example: Bearer
      expires_in:
        type: integer
        example: 1800
      permitted_resources:
        type: array
        items:
          type: string
        example:
        - flights:read
        - users:read:self
...
  Pagination:
    type: object
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total_pages:
        type: integer
      total_items:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      next_page:
        type: integer
        nullable: true
      prev_page:
        type: integer
        nullable: true
...
  UserListResponse:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/User'
      pagination:
        $ref: '#/definitions/Pagination'
...
paths:
  /users/token:
    post:
      summary: Issue scoped token
      description: Issue a short-lived JWT by passing an existing user ID. The token includes permitted_resources and is refreshed on successful protected API activity.
      operationId: issueUserToken
      tags:
      - Authentication
      parameters:
      - in: body
        name: body
        required: true
        schema:
          type: object
          required:
          - user_id
          properties:
            user_id:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        200:
          description: Token issued successfully
          schema:
            $ref: '#/definitions/UserResponse'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: &validation_error_example
              error: Bad Request
              message: Validation failed
              errors:
                field:
                - Validation error message
        404:
          description: User not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: &not_found_error_example
              error: Not Found
              message: Resource not found
  /users/{user_id}/token:
    get:
      summary: Issue scoped token by path user ID
      description: Issue a short-lived JWT for an existing user ID supplied in the URL.
      operationId: issueUserTokenById
      tags:
      - Authentication
      parameters:
      - name: user_id
        in: path
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: Token issued successfully
          schema:
            $ref: '#/definitions/UserResponse'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/NotFoundError'
          examples:
            application/json: *not_found_error_example
...
  /users/me:
    get:
      summary: Get token user profile
      description: Retrieve the profile information for the identity carried by the bearer token.
      operationId: getCurrentUser
      tags:
      - Users
      security:
      - Bearer: []
      responses:
        200:
          description: User profile details
          schema:
            type: object
            properties:
              user:
                $ref: '#/definitions/User'
        401:
          description: Invalid or missing token
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: &auth_error_example
              error: Unauthorized
              message: Invalid or missing authentication token
    patch:
      summary: Update token user profile
      description: Update the profile information for the identity carried by the bearer token.
      operationId: updateCurrentUser
      tags:
      - Users
      security:
      - Bearer: []
      parameters:
      - in: body
        name: body
        required: true
        schema:
          type: object
          properties:
            firstname:
              type: string
              maxLength: 30
            lastname:
              type: string
              maxLength: 30
            email:
              type: string
              format: email
      responses:
        200:
          description: Profile updated successfully
          schema:
            type: object
            properties:
              message:
                type: string
              user:
                $ref: '#/definitions/User'
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ValidationError'
          examples:
            application/json: *validation_error_example
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        409:
          description: Email already in use
          schema:
            $ref: '#/definitions/ConflictError'
          examples:
            application/json: &conflict_error_example
              error: Conflict
              message: Resource already exists or operation cannot be completed
  /users/:
    get:
      summary: List all users (Admin only)
      description: Get a paginated list of all users in the system. Only administrators can access this endpoint.
      operationId: listUsers
      tags:
      - Users
      security:
      - Bearer: []
      parameters:
      - $ref: '#/parameters/PageParam'
      - $ref: '#/parameters/PerPageParam'
      responses:
        200:
          description: Paginated list of all users
          schema:
            $ref: '#/definitions/UserListResponse'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/AuthError'
          examples:
            application/json: *auth_error_example
        403:
          description: Forbidden - Admin privileges required
          schema:
            $ref: '#/definitions/ForbiddenError'
          examples:
            application/json: &forbidden_error_example
              error: Forbidden
              message: You do not have permission to access this resource
...